+++ source/ftrack_connect/ui/widget/asset_manager.py
@@ -565,12 +565,6 @@ class AssetManagerWidget(QtWidgets.QWidget):
                 row, 5
             ).setCurrentIndex(newIndex)
 
-            newVersion = self.ui.AssertManagerTableWidget.cellWidget(
-                row, 5
-            ).currentText()
-
-            self.changeVersion(row, newVersion)
-
     def versionUpSelected(self):
         '''Version up selected assets.'''
         rows = self.getSelectedRows()
@@ -582,10 +576,6 @@ class AssetManagerWidget(QtWidgets.QWidget):
             self.ui.AssertManagerTableWidget.cellWidget(
                 row, 5
             ).setCurrentIndex(newIndex)
-            newVersion = self.ui.AssertManagerTableWidget.cellWidget(
-                row, 5
-            ).currentText()
-            self.changeVersion(row, newVersion)
 
     def versionLatestSelected(self):
         '''Version up assets to latest.'''
@@ -595,10 +585,6 @@ class AssetManagerWidget(QtWidgets.QWidget):
             self.ui.AssertManagerTableWidget.cellWidget(
                 row, 5
             ).setCurrentIndex(newIndex)
-            newVersion = self.ui.AssertManagerTableWidget.cellWidget(
-                row, 5
-            ).currentText()
-            self.changeVersion(row, newVersion)
 
     def selectAll(self):
         '''Select all assets rows.'''